requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0

# Email processing
email-validator>=2.0.0

# Database and API
supabase>=2.0.0
httpx>=0.24.0

# Web framework
streamlit>=1.28.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: For advanced features
# scikit-learn>=1.3.0  # For ML-based classification
# transformers>=4.30.0  # For fine-tuned models
# torch>=2.0.0  # For PyTorch-based models 